requirement.txt
asgiref==3.6.0
Django==4.1.6
django-jsonfield==1.4.1
djangorestframework==3.14.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7